Change random distribution type from int to ll

+++ 00_base_template/base.cpp
@@ -31,7 +31,7 @@ const int BASE = 31; // Hoặc 53, 313, 331,... (số nguyên tố)
 const int MAXN = 1e6 + 5; // Kích thước tối đa, thay đổi nếu cần
 mt19937 rng(chrono::high_resolution_clock::now().time_since_epoch().count());
 ll rand(ll left,ll right){
-    uniform_int_distribution<int> rd(left,right);
+    uniform_int_distribution<ll> rd(left,right);
     return rd(rng);
 }
 void fast_io() {
